FoodTrackerApi-main/requirements.txt
asgiref==3.7.2
django-cors-headers==4.1.0
Django
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.3.0
pymongo==3.12.3
pytz==2023.3
sqlparse==0.2.4
typing_extensions==4.7.0
tzdata==2023.3
